Fix when mse-set does not exist

+++ mse_generate.py
@@ -38,7 +38,8 @@ def run():
 	with open('build/set', 'w', encoding='UTF-8') as file:
 		file.write(output)
 	
-	os.remove('otakube.mse-set')
+	if os.path.exists('otakube.mse-set'):
+		os.remove('otakube.mse-set')
 	shutil.make_archive('otakube', 'zip', 'build')
 	os.rename('otakube.zip', 'otakube.mse-set')
